refactor(ControlledTextField): merge into a single TextField component
